+++ web/src/engine/websites/_index.ts
@@ -534,6 +534,7 @@ export { default as NocturneSummer } from './NocturneSummer';
 export { default as NoIndexScan } from './NoIndexScan';
 export { default as NoraNoFansub } from './NoraNoFansub';
 export { default as Noromax } from './Noromax';
+export { default as NoxManga } from './NoxManga';
 export { default as NyxScans } from './NyxScans';
 export { default as OlimpoScans } from './OlimpoScans';
 export { default as OlympusScanlation } from './OlympusScanlation';
@@ -594,7 +595,6 @@ export { default as ReadKomik } from './ReadKomik';
 export { default as ReadManga } from './ReadManga';
 export { default as ReadToon } from './ReadToon';
 export { default as Remanga } from './Remanga';
-export { default as Remangas } from './Remangas';
 export { default as ResetScans } from './ResetScans';
 export { default as Retsu } from './Retsu';
 export { default as Reyume } from './Reyume';
